fix: Separando los estilos del componente de prueba navbar y footer
Se separan los estilos inline de los componentes diseñados para prueba ;)  de el layout con contenido dinámico, ahora se encuentra separados en carpetas con su respectivos archivos module.scss
