+++ games-repository-frontend/src/Components/pruebas/footer/footer.module.scss
@@ -1,0 +1,35 @@
+.footer {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    gap: 1rem;
+    padding: 1rem;
+    background-color: #f5f5f5;
+    color: #333;
+    text-align: center;
+}
+  
+.column {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+  
+.menu {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+  
+.link {
+    text-decoration: none;
+    color: #333;
+    margin: 0;
+}
+  
+.button {
+    margin: 0.5rem 0;
+}
+
