+++ games-repository-frontend/src/Components/pruebas/navbar/navbar.module.scss
@@ -1,0 +1,20 @@
+nav{
+  background-color: #333;
+  padding: 1rem;
+}
+
+ul{
+  display: flex;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+li{
+  margin-right: 1rem;
+}
+
+a{
+  color: #fff;
+  text-decoration: none;
+}
